% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divModule.R
\name{divModule}
\alias{divModule}
\title{Dividing Modules.}
\usage{
divModule(case_union, control_union, deepSplit = 4, minModuleSize = 20)
}
\arguments{
\item{case_union}{a matrix of integrated diseased network.}

\item{control_union}{a matrix of integrated healthy network.}

\item{deepSplit}{For method "hybrid", can be either logical or integer in the range 0 to 4. For method "tree", must be logical. In both cases, provides a rough control over sensitivity to cluster splitting. The higher the value (or if TRUE), the more and smaller clusters will be produced. For the "hybrid" method, a finer control can be achieved via maxCoreScatter and minGap below.}

\item{minModuleSize}{Minimum module size.}
}
\value{
divided modules.
}
\description{
This is a function to divide integrated networks into different modules.
}
\examples{
Data(TestData)
divModule(TestData[[1]], TestData[[2]])

}
