% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netBuild.R
\name{netBuild}
\alias{netBuild}
\title{Building microbial networks}
\usage{
netBuild(
  case_dir,
  control_dir,
  net_case_dir,
  net_control_dir,
  method = "sparcc"
)
}
\arguments{
\item{case_dir}{string.The directory of diseased data set.}

\item{control_dir}{string.The directory of healthy data set.}

\item{net_case_dir}{string.The directory of network correlation of diseased data set.}

\item{net_control_dir}{string.The directory of network correlation of healthy data set.}

\item{method}{string. The method chosen to build microbial networks which can be "sparcc" or "pearson".}
}
\value{
microbial networks of case and control data set.
}
\description{
This is a function to build microbial networks using abundance matrix.
}
\examples{
case_dir = '/tests/case_dir'
control_dir = '/tests/control_dir'
net_case_dir = '/tests/net_case_dir'
net_control_dir = '/tests/net_control_dir'
netBuild(case_dir = case_dir,
     control_dir = control_dir,
     net_case_dir = net_case_dir,
     net_control_dir = net_control_dir,
     method = "sparcc")


}
