% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetMoss.R
\name{NetMoss}
\alias{NetMoss}
\title{Identifying Differential Bacteria Based on a Network Workflow.}
\usage{
NetMoss(
  case_dir,
  control_dir,
  net_case_dir,
  net_control_dir,
  deepSplit = 4,
  minModuleSize = 20
)
}
\arguments{
\item{case_dir}{string.The directory of diseased data set.}

\item{control_dir}{string.The directory of healthy data set.}

\item{net_case_dir}{string.The directory of network correlation of diseased data set.}

\item{net_control_dir}{string.The directory of network correlation of healthy data set.}

\item{deepSplit}{For method "hybrid", can be either logical or integer in the range 0 to 4. For method "tree", must be logical. In both cases, provides a rough control over sensitivity to cluster splitting. The higher the value (or if TRUE), the more and smaller clusters will be produced. For the "hybrid" method, a finer control can be achieved via maxCoreScatter and minGap below.}

\item{minModuleSize}{Minimum module size.}
}
\value{
NetMoss score and module division results of each bacterium.
}
\description{
This is a function to identify differential bacteria between case and control data set.
}
\examples{
case_dir = '/tests/case_dir'
control_dir = '/tests/control_dir'
net_case_dir = '/tests/net_case_dir'
net_control_dir = '/tests/net_control_dir'
NetMoss(case_dir = case_dir,
     control_dir = control_dir,
     net_case_dir = net_case_dir,
     net_control_dir = net_control_dir)

}
