% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netROC.R
\name{netROC}
\alias{netROC}
\title{Plot the combined ROC}
\usage{
netROC(case_dir, control_dir, marker, metadata, plot.roc = T, train.num = 20)
}
\arguments{
\item{case_dir}{string.The directory of diseased data set.}

\item{control_dir}{string.The directory of healthy data set.}

\item{marker}{the combined markers identified by NetMoss.}

\item{metadata}{the metadata of all input studies.}

\item{plot.roc}{logical.If TURE then the combined ROC will be plotted.}

\item{train.num}{numerical. training times for the module. 20 was set by default.}
}
\value{
a TPR and FPR table of the combined study.
}
\description{
This is a function to plot ROC of combined study.
}
\examples{
myROC = netROC(case_dir = case_dir,
     control_dir = control_dir,
     marker = marker,
     metadata = metadata,
     plot.roc = T,
     train.num = 20)

}
