% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetzGo.R
\name{NetzGO}
\alias{NetzGO}
\title{Calculating NetMoss Score}
\usage{
NetzGO(
  control_mat,
  case_mat,
  control_dist,
  case_dist,
  control_mod,
  case_mod,
  scaled
)
}
\arguments{
\item{control_mat}{a matrix of integrated healthy network.}

\item{case_mat}{a matrix of integrated diseased network.}

\item{control_dist}{a distance matrix among healthy models.}

\item{case_dist}{a distance matrix among diseased models.}

\item{control_mod}{module division of each microbe in healthy network.}

\item{case_mod}{module division of each microbe diseased network}

\item{scaled}{logical.If TURE then the NetMoss score are scaled.}
}
\value{
NetMoss score and module division results of each bacterium.
}
\description{
This is a function to calculate NetMoss score for each bacterium.
}
\examples{
data(TestData)
NetzGo(control_mat = TestData[[2]], case_mat = TestData[[1]],
    control_dist = TestData[[4]], case_dist = TestData[[3]],
    control_mod = TestData[[6]], case_mod = TestData[[5]], scaled = T)

}
